fix test

+++ TeachingRecordSystem/src/TeachingRecordSystem.AuthorizeAccess/Pages/RequestTrn/Identity.cshtml.cs
@@ -47,26 +47,34 @@ public class IdentityModel(AuthorizeAccessLinkGenerator linkGenerator, IFileServ
             ModelState.AddModelError(nameof(EvidenceFile), "Select a file");
         }
 
+
+        // Delete any previously uploaded file if they're uploading a new one,
+        // or choosing not to upload evidence (check for UploadEvidence != true because if
+        // UploadEvidence somehow got set to null we still want to delete the file)
+        if (EvidenceFileId.HasValue && EvidenceFile is not null)
+        {
+            await fileService.DeleteFileAsync(EvidenceFileId.Value);
+            EvidenceFileName = null;
+            EvidenceFileSizeDescription = null;
+            UploadedEvidenceFileUrl = null;
+            EvidenceFileId = null;
+        }
+
+        // Upload the file and set the display fields even if the rest of the form is invalid
+        // otherwise the user will have to re-upload every time they re-submit
+        if (EvidenceFile is not null)
+        {
+            using var stream = EvidenceFile.OpenReadStream();
+            var evidenceFileId = await fileService.UploadFileAsync(stream, EvidenceFile.ContentType);
+            EvidenceFileId = evidenceFileId;
+            EvidenceFileName = EvidenceFile?.FileName;
+            EvidenceFileSizeDescription = EvidenceFile?.Length.Bytes().Humanize();
+            UploadedEvidenceFileUrl = await fileService.GetFileUrlAsync(evidenceFileId, TimeSpan.FromMinutes(15));
+        }
+
         if (!ModelState.IsValid)
         {
             return this.PageWithErrors();
-        }
-
-        if (EvidenceFile is not null)
-        {
-            if (EvidenceFileId is not null)
-            {
-                await fileService.DeleteFileAsync(EvidenceFileId.Value);
-            }
-
-            using var stream = EvidenceFile.OpenReadStream();
-            var evidenceFileId = await fileService.UploadFileAsync(stream, EvidenceFile.ContentType);
-            await JourneyInstance!.UpdateStateAsync(state =>
-            {
-                state.EvidenceFileId = evidenceFileId;
-                state.EvidenceFileName = EvidenceFile.FileName;
-                state.EvidenceFileSizeDescription = EvidenceFile.Length.Bytes().Humanize();
-            });
         }
 
         return FromCheckAnswers == true ?
